Rewrote shader code and deleted unused stuff.

+++ src/Rendering/IRenderBuffer.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include <stdlib.h>
+#include <cstdlib>
 
 namespace mono
 {
@@ -12,7 +12,7 @@ namespace mono
         virtual ~IRenderBuffer() = default;
         virtual void UpdateData(const void* data, size_t offset, size_t count) = 0;
         virtual void Use() const = 0;
-        virtual size_t Id() const = 0;
+        virtual uint32_t Id() const = 0;
     };
 
     void ClearRenderBuffer();
